[COTLMPSERVER] Create a Data folder and move the Player class to that folder
Player is a data class, so move it to its appropriate place.
